test-cassandra/requirements.txt
cassandra_driver==3.25.0
spotipy==2.20.0
